--- HTML/settings.html
@@ -1,8 +1,0 @@
-<!DOCTYPE html>
-</head>
-    <title>設定</title>
-    <meta charset="utf-8">
-</head>
-<body>
-</body>
-</html>
