+++ nodejs/requestHandlers/index.js
@@ -1,0 +1,12 @@
+module.exports = {
+    topPage : require(__dirname + '/topPage'),
+    tvProgram : require(__dirname + '/tvProgram'),
+    viewTv : require(__dirname + '/viewTv'),
+    viewTvError : require(__dirname + '/viewTvError'),
+    responseSpecifiedFile : require(__dirname + '/responseSpecifiedFile'),
+    notFound : require(__dirname + '/notFound'),
+    epgrec : require(__dirname + '/epgrec'),
+    epgrecProgram : require(__dirname + '/epgrecProgram'),
+    epgrecRecorded : require(__dirname + '/epgrecRecorded'),
+    epgrecRecordedTag : require(__dirname + '/epgrecRecordedTag')
+}
